Refactor UserModel to support multiple branches
Moved UserModel from login to user feature and updated it to support multiple branches with a currentBranch property. Refactored all usages of branchId and branchName to use currentBranch.id and currentBranch.name. Updated related imports and UI to reflect the new structure, and added a Branch model. This change improves support for users associated with multiple branches.